move Config struct into its own package
